Version 0.1
Added SPK(Simple Python Keylogger) as well as a Batch-File which allows you to implement it inside the "StartUp" Folder in Windows OR use it on a Hard/Soft-Link to a Browser.

Keystrokes will be stored in the same folder such as the keylogger.pyw to change that or the filename, line 3 needs to be modified.

(Programm is running in quiet mode! Therefor needs to be CLOSED either through PC SHUTDOWN or KILLing it!!!)
~R